move cluster configuration to a separate file
